# Introduce object pooling