--- a/Jint.Benchmark/StopwatchBenchmark.cs
+++ b/Jint.Benchmark/StopwatchBenchmark.cs
@@ -1,9 +1,22 @@
 ﻿using BenchmarkDotNet.Attributes;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Diagnosers;
+using BenchmarkDotNet.Jobs;
 
 namespace Jint.Benchmark
 {
+    [Config(typeof(Config))]
     public class StopwatchBenchmark : SingleScriptBenchmark
     {
+        private class Config : ManualConfig
+        {
+            public Config()
+            {
+                Add(Job.ShortRun);
+                Add(MemoryDiagnoser.Default);
+            }
+        }
+ 
         [Params(1)]
         public override int N { get; set; }