--- a/Jint/Engine.cs
+++ b/Jint/Engine.cs
@@ -158,6 +158,7 @@ namespace Jint
             }
 
             ReferencePool = new ReferencePool();
+            CompletionPool = new CompletionPool();
 
             Eval = new EvalFunctionInstance(this, System.Array.Empty<string>(), LexicalEnvironment.NewDeclarativeEnvironment(this, ExecutionContext.LexicalEnvironment), StrictModeScope.IsStrictModeCode);
             Global.FastAddProperty("eval", Eval, true, false, true);
@@ -209,6 +210,7 @@ namespace Jint
         internal Options Options { get; private set; }
 
         internal ReferencePool ReferencePool { get; }
+        internal CompletionPool CompletionPool { get; }
 
         #region Debugger
         public delegate StepMode DebugStepDelegate(object sender, DebugInformation e);
@@ -339,11 +341,13 @@ namespace Jint
                 var result = _statements.ExecuteProgram(program);
                 if (result.Type == Completion.Throw)
                 {
-                    throw new JavaScriptException(result.GetValueOrDefault())
-                        .SetCallstack(this, result.Location);
+                    var ex = new JavaScriptException(result.GetValueOrDefault()).SetCallstack(this, result.Location);
+                    CompletionPool.Return(result);
+                    throw ex;
                 }
 
                 _completionValue = result.GetValueOrDefault();
+                CompletionPool.Return(result);
             }
 
             return this;
@@ -905,49 +909,6 @@ namespace Jint
                     }
                 }
             }
-        }
-    }
-
-    /// <summary>
-    /// Cache reusable Reference instances as we allocate them a lot.
-    /// </summary>
-    internal sealed class ReferencePool
-    {
-        private const int PoolSize = 10;
-        private readonly Reference[] _pool;
-        private int _currentSize;
-
-        public ReferencePool()
-        {
-            // pre-allocate so we don't show up in benchmarks
-            _pool = new Reference[PoolSize];
-            for (var i = 0; i < PoolSize; ++i)
-            {
-                _pool[i] = new Reference(JsValue.Undefined, string.Empty, false);
-            }
-
-            _currentSize = PoolSize;
-        }
-
-        public Reference Rent(JsValue baseValue, string name, bool strict)
-        {
-            if (_currentSize > 0)
-            {
-                _currentSize--;
-                var reference = _pool[_currentSize];
-                return reference.Reassign(baseValue, name, strict);
-            }
-            return new Reference(baseValue, name, strict);
-        }
-
-        public void Return(Reference reference)
-        {
-            if (reference == null || _currentSize >= PoolSize)
-            {
-                return;
-            }
-            _pool[_currentSize] = reference;
-            _currentSize++;
         }
     }
 }