--- a/Jint/Native/Object/ObjectInstance.cs
+++ b/Jint/Native/Object/ObjectInstance.cs
@@ -959,5 +959,11 @@ namespace Jint.Native.Object
 
             return false;
         }
+
+        internal virtual void Clear()
+        {
+            _intrinsicProperties?.Clear();
+            _properties?.Clear();
+        }
     }
 }