… Jint/Native/Array/ArrayPrototype.cs
… Jint/Native/Function/ScriptFunctionInstance.cs
… Jint/Pooling/ArgumentsInstancePool.cs
… Jint/Pooling/CompletionPool.cs
… Jint/Pooling/ObjectPool.cs
… Jint/Pooling/ReferencePool.cs
… Jint/Runtime/Completion.cs
… Jint/Runtime/Environments/DeclarativeEnvironmentRecord.cs
… Jint/Runtime/Environments/EnvironmentRecord.cs
… Jint/Runtime/Environments/LexicalEnvironment.cs
… Jint/Runtime/Environments/ObjectEnvironmentRecord.cs
… Jint/Runtime/ExpressionIntepreter.cs
… Jint/Runtime/References/Reference.cs
… Jint/Runtime/StatementInterpreter.cs